requirements.txt
langchain==0
langchain_community==0.0.16
langchain_core==0.1.17
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.3.29
beautifulsoup4==4.12.2
PyPDF2==3.0.1
huggingface-hub==0.14.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
mysql-connector-python==8.3.0
groq==0.4.2
langchain-groq==0.0.1
